test that complete files are served when shutdown is requested; wait up to 15s to finish *sending responses*, but immediately stop *receiving* requests including any in progress

+++ test/verm_spawner.rb
@@ -96,12 +96,20 @@ class VermSpawner
     end
     raise "Can't connect to our verm instance on #{host}"
   end
-  
-  def stop_verm
+
+  def request_stop
+    Process.kill('TERM', @verm_child_pid) if @verm_child_pid
+  end
+
+  def wait_for_stop
     return unless @verm_child_pid
-    Process.kill('TERM', @verm_child_pid)
     Process.wait(@verm_child_pid)
     raise "process terminated unsuccessfully: #{$?.inspect}" unless $?.success?
     @verm_child_pid = nil
+  end
+
+  def stop_verm
+    request_stop
+    wait_for_stop
   end
 end
